# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "YouTube Proxy API",
    "version": "1.0.0",
    "description": "Store YouTube URLs + hashes and fetch YouTube metadata"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/videos": {
      "post": {
        "summary": "Create URL-hash mapping",
        "tags": [
          "Videos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "hash": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 128
                  },
                  "userId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "summary": "List stored videos with filtering and sorting",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "userId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "tags",
            "required": false,
            "description": "Comma separated tag list (from metadata.extra.tags)."
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "any",
                "all"
              ]
            },
            "in": "query",
            "name": "tagsMode",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "minElo",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "maxElo",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "elo",
                "created",
                "category",
                "updated"
              ]
            },
            "in": "query",
            "name": "sortBy",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "in": "query",
            "name": "sortOrder",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/videos/{hash}": {
      "get": {
        "summary": "Get stored mapping by hash",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "summary": "Update service-managed video fields (userId, elo, category)",
        "tags": [
          "Videos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "elo": {
                    "type": "number"
                  },
                  "category": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/videos/{hash}/metadata": {
      "get": {
        "summary": "Get full video metadata",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "refresh",
            "required": false,
            "description": "Set to true to force refresh from YouTube (for example: true, 1, yes)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/videos/{hash}/thumbnail": {
      "get": {
        "summary": "Get video thumbnail URLs",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "refresh",
            "required": false,
            "description": "Set to true to force refresh from YouTube (for example: true, 1, yes)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/videos/{hash}/stats": {
      "get": {
        "summary": "Get video statistics",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "refresh",
            "required": false,
            "description": "Set to true to force refresh from YouTube (for example: true, 1, yes)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/videos/{hash}/channel": {
      "get": {
        "summary": "Get channel metadata for a video",
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "refresh",
            "required": false,
            "description": "Set to true to force refresh from YouTube (for example: true, 1, yes)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}